Move fetch versions to autocomplete folder
