+++ src/client/common/helpers/fetch/autocomplete/index.js
@@ -1,0 +1,3 @@
+import { fetchVersions } from '~/src/client/common/helpers/fetch/autocomplete/fetch-versions'
+
+export { fetchVersions }
