+++ src/client/common/helpers/fetch/select/index.js
@@ -1,4 +1,3 @@
 import { fetchMemory } from '~/src/client/common/helpers/fetch/select/fetch-memory'
-import { fetchVersions } from '~/src/client/common/helpers/fetch/select/fetch-versions'
 
-export { fetchMemory, fetchVersions }
+export { fetchMemory }
